Sanitize input distribution list
I ran into this problem while trying to test locally without any of the
env vars setup. The current code was passing a slice with an empty
string to the plugin. And this was then replacing the distribution list
and effectively making it empty, from the POV of the command.